requirements.txt
asyncpg
sqlalchemy
pydantic
pydantic-settings
fastapi
uvicorn
docker
python-dotenv
redis
pyjwt
cryptography
bcrypt
python-multipart
pydantic[email]